+++ src/pairwise/alignment.jl
@@ -14,6 +14,21 @@ mutable struct PairwiseAlignment{S1,S2}
     b::S2
 end
 
+# Getter functions
+"""
+    alignment(pairwise_alignment)
+
+Gets the underlying [`Alignment`](@ref) from `pairwise_alignment`.
+"""
+alignment(aln::PairwiseAlignment) = alignment(aln.a)
+
+"""
+    sequence(pairwise_alignment)
+
+Gets the query sequence of `pairwise_alignment`.
+"""
+sequence(aln::PairwiseAlignment) = sequence(aln.a)
+
 function Base.iterate(aln::PairwiseAlignment, ij=(2,1))
     i, j = ij
     if i > lastindex(aln.a.aln.anchors)