+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Getter functions for `AlignedSequence`, `PairwiseAlignment`, and
+  `PairwiseAlignmentResult`
+
 ## [2.2.0]
 
 ### Added